+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVVFS9S8GHVJJ32RVZC85KD2"
+        "01JVVX9GVWQHBAXPEZMEVTD9C7"
     )
     r"""The id of the resource"""