+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVVFS9S6VJ9S61HP1P9TY5V3"
+        "01JVVX9GVVQ0E7SE1F78ER9H5B"
     )
     r"""The id of the resource"""